Adds real-time audio transcription via Vosk
Integrates Vosk ASR to provide real-time transcription of the transmitted audio.
The client now connects to a WebSocket server to receive a unique identifier.
This identifier is then included in the header of each audio chunk sent to the UDP server, enabling WebSocket association.
This allows the server to transcript audio chunks by client and send trough Websocket for client specific transcription.
The transmitted audio includes a WebSocket ID in its header, and handles WebSocket connections for client identification.